First draft layout complete
All elements arranged satisfactorily. Was overcomplicating the structure trying to use grid-template-areas.
Need to do a deep dive on the subject as I am still not entirely sure why. In any case, it was unnecessary.
The desired look was achieved through setting the 'label' and 'input' elements to 'flex' and the
'flex-direction' to column. Once that hurdle was overcome, they rest was a relatively straightforward, less
brow furrowing experience.